# Speed up parsing