--- a/parser_test.go
+++ b/parser_test.go
@@ -29,10 +29,6 @@ var messageTests = []struct {
 		IsNil: true,
 	},
 	{
-		Expect: ":asd  :",
-		IsNil:  true,
-	},
-	{
 		Expect: ":A",
 		IsNil:  true,
 	},