```
# Before
~/g/s/g/b/irc % go test -bench=. -benchmem
BenchmarkParseMessage-4      5000000           307 ns/op          85 B/op          2 allocs/op
BenchmarkParsePrefix-4       5000000           339 ns/op         112 B/op          3 allocs/op
PASS
ok      github.com/belak/irc    3.917s

# After
~/g/s/g/b/irc % go test -bench=. -benchmem
BenchmarkParseMessage-4     10000000           151 ns/op          68 B/op          1 allocs/op
BenchmarkParsePrefix-4      20000000            97.5 ns/op        48 B/op          1 allocs/op
PASS
ok      github.com/belak/irc    3.724s
```

The biggest win is in ParsePrefix... I'm not sure how much value we'll get out of this, as message parsing is still pretty fast.